# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "surplus crudAPI test API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "surplus crudAPI test API Documentation"
        }
    ],
    "paths": {
        "/api": {},
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "get categories list",
                "description": "Return list of categories",
                "operationId": "ead5fb1cf2a7453acb2e37928a1b1eaf",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Post category data",
                "description": "store new category data",
                "operationId": "8cdcc1c8ebab5e0be2ba6d4ac19969b7",
                "requestBody": {
                    "description": "Pass category data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "enable"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Paijo"
                                    },
                                    "enable": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data Category berhasil disimpan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Categories",
            "description": "API Endpoints of Categories"
        },
        {
            "name": "Products",
            "description": "API Endpoints of Products"
        }
    ]
}